# clientes-cuentas-microservicio-infrastructure/src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: clientes-cuentas-microservicio
  description: Aplicacion Gestion Clientes & Cuentas Bancarias
  version: 1.0.0
servers:
  - url: http://localhost:8080/{basepath}
    variables:
      basepath:
        default: api/v1
tags:
  - name: Clientes
    description: Operaciones relacionadas con clientes
  - name: Cuentas
    description: Operaciones relacionadas con cuentas bancarias
  - name: Movimientos
    description: Operaciones relacionadas con movimientos bancarios
paths:
  /clientes:
    get:
      tags:
        - Clientes
      summary: Obtener todos los clientes
      description: Consulta para recuperar todos los clientes
      operationId: findAllClientes
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClienteRS'
        '204':
          description: 'Clientes no encontrados'
        '400':
          description: 'Bad Request'
        '500':
          description: 'ERROR GET Clientes: Internal Server ERROR'
  /clientes/adultos:
    get:
      tags:
        - Clientes
      summary: Obtener los clientes mayores de edad
      description: Consulta para recuperar todos los clientes mayores de edad
      operationId: findClientesAdultos
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClienteRS'
        '204':
          description: 'Clientes no encontrados'
        '400':
          description: 'Bad Request'
        '500':
          description: 'ERROR GET Clientes Adultos: Internal Server ERROR'
  /clientes/total/{cantidad}:
    get:
      tags:
        - Clientes
      summary: Obtener los clientes con una suma total de sus cuentas superior a la cantidad dada
      description: Consulta para recuperar los clientes con una suma total de sus cuentas superior a la cantidad dada
      operationId: findClientesByTotalGreaterThan
      parameters:
        - name: cantidad
          in: path
          required: true
          schema:
            type: number
            format: float
          description: Cantidad mínima
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClienteRS'
        '204':
          description: 'Clientes no encontrados'
        '400':
          description: 'Bad Request'
        '500':
          description: 'ERROR GET getClientesCantidadSuperior: Internal Server ERROR'
  /clientes/{dni}:
    get:
      tags:
        - Clientes
      summary: Obtener Cliente By DNI
      description: Obtiene un Cliente en base a un DNI introducido
      operationId: findClienteByDni
      parameters:
        - name: dni
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9]{8}[A-Za-z]$"
          description: DNI español
      responses:
        '200':
          description: 'Cliente encontrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteRS'
        '204':
          description: 'Cliente no encontrado'
        '400':
          description: 'Bad Request'
        '500':
          description: 'ERROR GET getClienteById: Internal Server ERROR'
  /cuentas:
    post:
      tags:
        - Cuentas
      summary: Dar de alta una cuenta a un cliente asociado
      description: Dar de alta una cuenta a un cliente asociado
      operationId: createCuenta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CuentaRQ'
      responses:
        '201':
          description: 'Cuenta Creada Correctamente'
        '400':
          description: 'Bad Request'
        '500':
          description: "ERROR POST postAddCuenta: Internal Server ERROR"

  /cuentas/{idCuenta}:
    put:
      tags:
        - Cuentas
      summary: Actualizar saldo cuenta
      description: Actualiza el total de una cuenta dada
      operationId: updateCuenta
      parameters:
        - name: idCuenta
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Identificador Cuenta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CuentaUpdateRQ'
      responses:
        '200':
          description: 'Cuenta actualizada correctamente'
        '400':
          description: 'Bad Request'
        '500':
          description: "ERROR PUT putUpdateTotalCuenta: Internal Server ERROR"
  /movimientos/{idCuenta}:
    get:
      tags:
        - Movimientos
      summary: Obtener movimientos de una cuenta
      description: Consulta para recuperar los movimientos de una cuenta dada
      operationId: findMovimientosByCuentaId
      parameters:
        - name: idCuenta
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: "Identificador Cuenta"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
          description: "Paginación - número de página (0-indexada)"
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
          description: "Paginación - tamaño de página"
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: "desc"
          description: "Paginación - ordenación"
        - name: tipo
          in: query
          required: false
          schema:
            type: string
            enum:
              - INGRESO
              - RETIRADA
              - TRANSFERENCIA
          description: "Filtro opcional por tipo de movimiento (INGRESO, RETIRADA o TRANSFERENCIA)"
      responses:
        '200':
          description: Lista de movimientos de la cuenta
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/MovimientoPageRS'
        '204':
          description: 'Movimientos no encontrados'
        '400':
          description: 'Bad Request'
        '500':
          description: "ERROR GET getMovimientosByCuenta: Internal Server ERROR"
components:
  schemas:
    CuentaRQ:
      type: object
      required:
        - dniCliente
        - tipoCuenta
        - total
      properties:
        dniCliente:
          type: string
          pattern: "^[0-9]{8}[A-Za-z]$"
          description: DNI español
        nombre:
          type: string
        apellido1:
          type: string
        apellido2:
          type: string
        tipoCuenta:
          type: string
          enum:
            - NORMAL
            - JUNIOR
            - PREMIUM
          description: Tipo de cuenta
        total:
          type: number
          format: float
          description: Saldo inicial de la cuenta
    CuentaUpdateRQ:
      type: object
      required:
        - total
      properties:
        total:
          type: number
          format: float
          description: Saldo de la cuenta
    ClienteRS:
      type: object
      required:
        - dniCliente
      properties:
        dniCliente:
          type: string
          pattern: "^[0-9]{8}[A-Za-z]$"
          description: DNI español
        nombre:
          type: string
        apellido1:
          type: string
        apellido2:
          type: string
        fechaNacimiento:
          type: string
          format: date
          description: Fecha de nacimiento del cliente (YYYY-MM-DD)
        cuentas:
          type: array
          items:
            $ref: '#/components/schemas/CuentaRS'
    CuentaRS:
      type: object
      required:
        - id
        - tipoCuenta
        - total
      properties:
        id:
          type: integer
          format: int64
        tipoCuenta:
          type: string
          description: Tipo de cuenta
        total:
          type: number
          format: double
          description: Saldo de la cuenta
    MovimientoPageRS:
        type: object
        required:
            - content
            - page
            - size
            - totalElements
            - totalPages
        properties:
            content:
              type: array
              items:
                  $ref: '#/components/schemas/MovimientoRS'
            page:
              type: integer
              format: int32
              description: Número de página actual (0-indexada)
            size:
              type: integer
              format: int32
              description: Tamaño de página solicitado
            totalElements:
              type: integer
              format: int64
              description: Número total de elementos disponibles
            totalPages:
              type: integer
              format: int32
              description: Número total de páginas disponibles
    MovimientoRS:
      type: object
      required:
        - id
        - cuentaId
        - fechaOperacion
        - tipoMovimiento
        - importe
      properties:
        id:
          type: integer
          format: int64
        cuentaId:
          type: integer
          format: int64
        fechaOperacion:
          type: string
          format: date-time
          description: Fecha y hora del movimiento (YYYY-MM-DDTHH:mm:ss)
        tipoMovimiento:
          type: string
          enum:
            - INGRESO
            - RETIRADA
            - TRANSFERENCIA
          description: Tipo de movimiento (INGRESO, RETIRADA o TRANSFERENCIA)
        importe:
          type: number
          format: double
          description: Cantidad del movimiento
        descripcion:
          type: string
          description: Descripción del movimiento (opcional)